+++ webapp/src/test/java/ar/edu/itba/paw/webapp/controller/UserBanAppealControllerTest.java
@@ -102,7 +102,7 @@ class UserBanAppealControllerTest {
                 null,
                 null,
                 null,
-                0,
+                (short) 0,
                 null,
                 null,
                 null,
