+++ webapp/src/test/java/ar/edu/itba/paw/webapp/security/BannedAccountAuthorizationFilterTest.java
@@ -81,7 +81,7 @@ class BannedAccountAuthorizationFilterTest {
                 null,
                 null,
                 null,
-                0,
+                (short) 0,
                 null,
                 null,
                 null,
